Refactor: Move to new package
