+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/authorized_role/security_manager/SecurityManagerView.java
@@ -18,8 +18,8 @@
 package bisq.desktop.main.content.authorized_role.security_manager;
 
 import bisq.bonded_roles.bonded_role.BondedRole;
-import bisq.bonded_roles.security_manager.AlertType;
-import bisq.bonded_roles.security_manager.AuthorizedAlertData;
+import bisq.bonded_roles.security_manager.alert.AlertType;
+import bisq.bonded_roles.security_manager.alert.AuthorizedAlertData;
 import bisq.desktop.common.view.View;
 import bisq.desktop.components.controls.AutoCompleteComboBox;
 import bisq.desktop.components.controls.MaterialTextArea;
