+++ bonded-roles/src/main/java/bisq/bonded_roles/security_manager/alert/AlertService.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.bonded_roles.security_manager;
+package bisq.bonded_roles.security_manager.alert;
 
 import bisq.bonded_roles.BondedRoleType;
 import bisq.bonded_roles.bonded_role.AuthorizedBondedRolesService;
