+++ bonded-roles/src/main/java/bisq/bonded_roles/security_manager/alert/AlertType.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.bonded_roles.security_manager;
+package bisq.bonded_roles.security_manager.alert;
 
 import bisq.common.proto.ProtoEnum;
 import bisq.common.util.ProtobufUtils;
